Add entities using the JDL model `Simphonyv2`

+++ src/main/java/mc/sbm/simphonycloud/domain/BarCodeTrans.java
@@ -20,7 +20,7 @@ public class BarCodeTrans implements Serializable {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "sequenceGenerator")
     @SequenceGenerator(name = "sequenceGenerator")
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "loc_ref")
     private String locRef;
@@ -51,16 +51,16 @@ public class BarCodeTrans implements Serializable {
 
     // jhipster-needle-entity-add-field - JHipster will add fields here
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public BarCodeTrans id(Long id) {
+    public BarCodeTrans id(Integer id) {
         this.setId(id);
         return this;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
