+++ src/main/java/mc/sbm/simphonycloud/domain/CheckTrans.java
@@ -21,7 +21,7 @@ public class CheckTrans implements Serializable {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "sequenceGenerator")
     @SequenceGenerator(name = "sequenceGenerator")
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "rvc_ref")
     private Integer rvcRef;
@@ -94,16 +94,16 @@ public class CheckTrans implements Serializable {
 
     // jhipster-needle-entity-add-field - JHipster will add fields here
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public CheckTrans id(Long id) {
+    public CheckTrans id(Integer id) {
         this.setId(id);
         return this;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
