+++ src/main/java/mc/sbm/simphonycloud/domain/CodeRaisonBI.java
@@ -21,7 +21,7 @@ public class CodeRaisonBI implements Serializable {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "sequenceGenerator")
     @SequenceGenerator(name = "sequenceGenerator")
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @NotNull
     @Column(name = "nom_court", nullable = false)
@@ -42,16 +42,16 @@ public class CodeRaisonBI implements Serializable {
 
     // jhipster-needle-entity-add-field - JHipster will add fields here
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public CodeRaisonBI id(Long id) {
+    public CodeRaisonBI id(Integer id) {
         this.setId(id);
         return this;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
