+++ src/main/java/mc/sbm/simphonycloud/domain/FamilyGroupCnC.java
@@ -20,7 +20,7 @@ public class FamilyGroupCnC implements Serializable {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "sequenceGenerator")
     @SequenceGenerator(name = "sequenceGenerator")
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "nom")
     private String nom;
@@ -33,16 +33,16 @@ public class FamilyGroupCnC implements Serializable {
 
     // jhipster-needle-entity-add-field - JHipster will add fields here
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public FamilyGroupCnC id(Long id) {
+    public FamilyGroupCnC id(Integer id) {
         this.setId(id);
         return this;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
