+++ src/main/java/mc/sbm/simphonycloud/domain/GuestCheckBI.java
@@ -25,7 +25,7 @@ public class GuestCheckBI implements Serializable {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "sequenceGenerator")
     @SequenceGenerator(name = "sequenceGenerator")
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "guest_check_id")
     private Integer guestCheckId;
@@ -91,16 +91,16 @@ public class GuestCheckBI implements Serializable {
 
     // jhipster-needle-entity-add-field - JHipster will add fields here
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public GuestCheckBI id(Long id) {
+    public GuestCheckBI id(Integer id) {
         this.setId(id);
         return this;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
