+++ src/main/java/mc/sbm/simphonycloud/domain/LocationCnC.java
@@ -20,7 +20,7 @@ public class LocationCnC implements Serializable {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "sequenceGenerator")
     @SequenceGenerator(name = "sequenceGenerator")
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "hier_unit_id")
     private Integer hierUnitId;
@@ -105,16 +105,16 @@ public class LocationCnC implements Serializable {
 
     // jhipster-needle-entity-add-field - JHipster will add fields here
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public LocationCnC id(Long id) {
+    public LocationCnC id(Integer id) {
         this.setId(id);
         return this;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
