+++ src/main/java/mc/sbm/simphonycloud/domain/OrderChannelBI.java
@@ -21,7 +21,7 @@ public class OrderChannelBI implements Serializable {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "sequenceGenerator")
     @SequenceGenerator(name = "sequenceGenerator")
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "num")
     private Integer num;
@@ -41,16 +41,16 @@ public class OrderChannelBI implements Serializable {
 
     // jhipster-needle-entity-add-field - JHipster will add fields here
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public OrderChannelBI id(Long id) {
+    public OrderChannelBI id(Integer id) {
         this.setId(id);
         return this;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
