+++ src/main/java/mc/sbm/simphonycloud/domain/PointDeVenteCnC.java
@@ -20,7 +20,7 @@ public class PointDeVenteCnC implements Serializable {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "sequenceGenerator")
     @SequenceGenerator(name = "sequenceGenerator")
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "loc_hier_unit_id")
     private Integer locHierUnitId;
@@ -48,16 +48,16 @@ public class PointDeVenteCnC implements Serializable {
 
     // jhipster-needle-entity-add-field - JHipster will add fields here
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public PointDeVenteCnC id(Long id) {
+    public PointDeVenteCnC id(Integer id) {
         this.setId(id);
         return this;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
