+++ src/main/java/mc/sbm/simphonycloud/domain/PointDeVenteTrans.java
@@ -20,7 +20,7 @@ public class PointDeVenteTrans implements Serializable {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "sequenceGenerator")
     @SequenceGenerator(name = "sequenceGenerator")
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "rvc_ref")
     private Integer rvcRef;
@@ -39,16 +39,16 @@ public class PointDeVenteTrans implements Serializable {
 
     // jhipster-needle-entity-add-field - JHipster will add fields here
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public PointDeVenteTrans id(Long id) {
+    public PointDeVenteTrans id(Integer id) {
         this.setId(id);
         return this;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
