+++ src/main/java/mc/sbm/simphonycloud/domain/RemiseBI.java
@@ -20,7 +20,7 @@ public class RemiseBI implements Serializable {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "sequenceGenerator")
     @SequenceGenerator(name = "sequenceGenerator")
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "num")
     private Integer num;
@@ -42,16 +42,16 @@ public class RemiseBI implements Serializable {
 
     // jhipster-needle-entity-add-field - JHipster will add fields here
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public RemiseBI id(Long id) {
+    public RemiseBI id(Integer id) {
         this.setId(id);
         return this;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
