+++ src/main/java/mc/sbm/simphonycloud/domain/RemiseTrans.java
@@ -20,7 +20,7 @@ public class RemiseTrans implements Serializable {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "sequenceGenerator")
     @SequenceGenerator(name = "sequenceGenerator")
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "org_short_name")
     private String orgShortName;
@@ -48,16 +48,16 @@ public class RemiseTrans implements Serializable {
 
     // jhipster-needle-entity-add-field - JHipster will add fields here
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public RemiseTrans id(Long id) {
+    public RemiseTrans id(Integer id) {
         this.setId(id);
         return this;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
