+++ src/main/java/mc/sbm/simphonycloud/domain/SocieteTrans.java
@@ -20,7 +20,7 @@ public class SocieteTrans implements Serializable {
     @GeneratedValue(strategy = GenerationType.SEQUENCE, generator = "sequenceGenerator")
     @SequenceGenerator(name = "sequenceGenerator")
     @Column(name = "id")
-    private Long id;
+    private Integer id;
 
     @Column(name = "nom")
     private String nom;
@@ -30,16 +30,16 @@ public class SocieteTrans implements Serializable {
 
     // jhipster-needle-entity-add-field - JHipster will add fields here
 
-    public Long getId() {
+    public Integer getId() {
         return this.id;
     }
 
-    public SocieteTrans id(Long id) {
+    public SocieteTrans id(Integer id) {
         this.setId(id);
         return this;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
