+++ src/main/java/mc/sbm/simphonycloud/repository/CommissionServiceTransRepository.java
@@ -9,4 +9,4 @@ import org.springframework.stereotype.Repository;
  */
 @SuppressWarnings("unused")
 @Repository
-public interface CommissionServiceTransRepository extends JpaRepository<CommissionServiceTrans, Long> {}
+public interface CommissionServiceTransRepository extends JpaRepository<CommissionServiceTrans, Integer> {}
